+++ cropclassification/preprocess/_prepare_input_BEFL.py
@@ -55,6 +55,14 @@ columns_BEFL_to_keep = [
 ndvi_latecrop_count = "latecrop_ndvi_count"
 ndvi_latecrop_median = "latecrop_ndvi_median"
 
+# Set parcels having a main crop that stays late on the field to another class, as
+# the main crop will still be on the field:
+# TODO: should be in REFE instead of hardcoded!!!
+late_main_crops = {
+    "MON_CG_BIETEN": ["71", "91", "8532", "9532"],
+    "MON_CG_MAIS_OF_MENGTEELT": ["201", "202", "396"],
+}
+
 
 def prepare_input(
     input_parcel_path: Path,
@@ -319,49 +327,36 @@ def prepare_input(
             min_parcels_in_class=min_parcels_in_class,
             is_groundtruth=True,
         )
-    elif classtype_to_prepare == "LATECROP-EARLY":
-        parceldata_df = prepare_input_latecrop_early(
-            parceldata_df=parceldata_df,
-            column_BEFL_latecrop=column_BEFL_latecrop,
-            column_BEFL_latecrop2=column_BEFL_latecrop2,
-            column_BEFL_maincrop=column_BEFL_crop,
-            column_output_class=conf.columns["class"],
-            classes_refe_path=classes_refe_path,
-            min_parcels_in_class=min_parcels_in_class,
-            is_groundtruth=False,
-        )
-    elif classtype_to_prepare == "LATECROP":
+    elif classtype_to_prepare.startswith("LATECROP"):
+        if classtype_to_prepare.startswith("LATECROP-EARLY"):
+            scope = "EARLY_MAINCROP"
+        elif classtype_to_prepare.startswith("LATECROP-LATE"):
+            scope = "LATE_MAINCROP"
+        elif classtype_to_prepare.startswith("LATECROP-ALL"):
+            scope = "ALL"
+        else:
+            raise ValueError(f"Unknown late crop scope: {classtype_to_prepare}")
+
+        is_groundtruth = classtype_to_prepare.endswith("-GROUNDTRUTH")
+        if not is_groundtruth:
+            column_latecrop = column_BEFL_latecrop
+            column_latecrop2 = column_BEFL_latecrop2
+            column_output_class = conf.columns["class"]
+        else:
+            column_latecrop = column_BEFL_latecrop_gt_verified
+            column_latecrop2 = None
+            column_output_class = conf.columns["class_groundtruth"]
+
         parceldata_df = prepare_input_latecrop(
             parceldata_df=parceldata_df,
-            column_BEFL_latecrop=column_BEFL_latecrop,
-            column_BEFL_latecrop2=column_BEFL_latecrop2,
+            column_BEFL_latecrop=column_latecrop,
+            column_BEFL_latecrop2=column_latecrop2,
             column_BEFL_maincrop=column_BEFL_crop,
-            column_output_class=conf.columns["class"],
+            column_output_class=column_output_class,
             classes_refe_path=classes_refe_path,
             min_parcels_in_class=min_parcels_in_class,
-            is_groundtruth=False,
-        )
-    elif classtype_to_prepare == "LATECROP-EARLY-GROUNDTRUTH":
-        parceldata_df = prepare_input_latecrop_early(
-            parceldata_df=parceldata_df,
-            column_BEFL_latecrop=column_BEFL_latecrop_gt_verified,
-            column_BEFL_latecrop2=None,
-            column_BEFL_maincrop=column_BEFL_crop,
-            column_output_class=conf.columns["class_groundtruth"],
-            classes_refe_path=classes_refe_path,
-            min_parcels_in_class=min_parcels_in_class,
-            is_groundtruth=True,
-        )
-    elif classtype_to_prepare == "LATECROP-GROUNDTRUTH":
-        parceldata_df = prepare_input_latecrop(
-            parceldata_df=parceldata_df,
-            column_BEFL_latecrop=column_BEFL_latecrop_gt_verified,
-            column_BEFL_latecrop2=None,
-            column_BEFL_maincrop=column_BEFL_crop,
-            column_output_class=conf.columns["class_groundtruth"],
-            classes_refe_path=classes_refe_path,
-            min_parcels_in_class=min_parcels_in_class,
-            is_groundtruth=True,
+            is_groundtruth=is_groundtruth,
+            scope=scope,
         )
     elif classtype_to_prepare == "FABACEAE":
         parceldata_df = prepare_input_fabaceae(
@@ -382,6 +377,58 @@ def prepare_input(
         )
     elif classtype_to_prepare == "FABACEAE-GROUNDTRUTH":
         parceldata_df = prepare_input_fabaceae(
+            parceldata_df=parceldata_df,
+            column_BEFL_cropcode=column_BEFL_crop_gt_verified,
+            column_output_class=conf.columns["class_groundtruth"],
+            classes_refe_path=classes_refe_path,
+            min_parcels_in_class=min_parcels_in_class,
+            is_groundtruth=True,
+        )
+    elif classtype_to_prepare == "RUGGENTEELT":
+        parceldata_df = prepare_input_ruggenteelt(
+            parceldata_df=parceldata_df,
+            column_BEFL_cropcode=column_BEFL_crop_declared,
+            column_output_class=conf.columns["class_declared"],
+            classes_refe_path=classes_refe_path,
+            min_parcels_in_class=1,
+            is_groundtruth=False,
+        )
+        parceldata_df = prepare_input_ruggenteelt(
+            parceldata_df=parceldata_df,
+            column_BEFL_cropcode=column_BEFL_crop,
+            column_output_class=conf.columns["class"],
+            classes_refe_path=classes_refe_path,
+            min_parcels_in_class=min_parcels_in_class,
+            is_groundtruth=False,
+        )
+    elif classtype_to_prepare == "RUGGENTEELT-GROUNDTRUTH":
+        parceldata_df = prepare_input_ruggenteelt(
+            parceldata_df=parceldata_df,
+            column_BEFL_cropcode=column_BEFL_crop_gt_verified,
+            column_output_class=conf.columns["class_groundtruth"],
+            classes_refe_path=classes_refe_path,
+            min_parcels_in_class=min_parcels_in_class,
+            is_groundtruth=True,
+        )
+    elif classtype_to_prepare == "RUGGENTEELT-EARLY":
+        parceldata_df = prepare_input_ruggenteelt_early(
+            parceldata_df=parceldata_df,
+            column_BEFL_cropcode=column_BEFL_crop_declared,
+            column_output_class=conf.columns["class_declared"],
+            classes_refe_path=classes_refe_path,
+            min_parcels_in_class=1,
+            is_groundtruth=False,
+        )
+        parceldata_df = prepare_input_ruggenteelt_early(
+            parceldata_df=parceldata_df,
+            column_BEFL_cropcode=column_BEFL_crop,
+            column_output_class=conf.columns["class"],
+            classes_refe_path=classes_refe_path,
+            min_parcels_in_class=min_parcels_in_class,
+            is_groundtruth=False,
+        )
+    elif classtype_to_prepare == "RUGGENTEELT-EARLY-GROUNDTRUTH":
+        parceldata_df = prepare_input_ruggenteelt_early(
             parceldata_df=parceldata_df,
             column_BEFL_cropcode=column_BEFL_crop_gt_verified,
             column_output_class=conf.columns["class_groundtruth"],
@@ -1094,6 +1141,7 @@ def prepare_input_latecrop(
     classes_refe_path: Path,
     min_parcels_in_class: int,
     is_groundtruth: bool,
+    scope: str,
 ):
     """Prepare input file for use in the latecrop marker.
 
@@ -1108,7 +1156,9 @@ def prepare_input_latecrop(
     # Check if parameters are OK and init some extra params
     # -----------------------------------------------------
     if not classes_refe_path.exists():
-        raise Exception(f"Input classes file doesn't exist: {classes_refe_path}")
+        raise ValueError(f"Input classes file doesn't exist: {classes_refe_path}")
+    if scope is None or scope not in ["ALL", "EARLY_MAINCROP", "LATE_MAINCROP"]:
+        raise ValueError(f"Invalid value for scope: {scope}")
 
     # Convert the crops to unicode, in case the input is int
     if column_BEFL_latecrop in parceldata_df.columns:
@@ -1153,7 +1203,7 @@ def prepare_input_latecrop(
                 conf.columns["class"],
                 ndvi_latecrop_count,
                 ndvi_latecrop_median,
-                "MON_CROPGROUP",
+                "MON_CG",
                 "IS_PERM_BEDEKKING",
             ]
             and column not in columns_BEFL_to_keep
@@ -1210,6 +1260,40 @@ def prepare_input_latecrop(
             value=parceldata_df[f"{column_output_class_orig}_LATECROP2"],
         )
 
+    # If limited scope, apply it
+    if scope == "EARLY_MAINCROP":
+        # Only process parcels with an early main crop.
+        # Hence, set parcels with a main crop that stays late on the field to a class
+        # that can be ignored later on.
+        for class_name, cropcodes in late_main_crops.items():
+            parceldata_df.loc[
+                parceldata_df[column_BEFL_maincrop].isin(cropcodes), column_output_class
+            ] = class_name
+            if not is_groundtruth:
+                parceldata_df.loc[
+                    parceldata_df[column_BEFL_maincrop].isin(cropcodes),
+                    conf.columns["class_declared"],
+                ] = class_name
+
+    elif scope == "LATE_MAINCROP":
+        # Only process parcels with a late main crop.
+        # Hence, set parcels with a main crop that is remove early to an IGNORE class.
+        early_maincrop_classname = "IGNORE:EARLY_MAINCROP"
+
+        early_maincrops = []
+        for _, cropcodes in late_main_crops.items():
+            early_maincrops.extend(cropcodes)
+
+        parceldata_df.loc[
+            ~parceldata_df[column_BEFL_maincrop].isin(early_maincrops),
+            column_output_class,
+        ] = early_maincrop_classname
+        if not is_groundtruth:
+            parceldata_df.loc[
+                ~parceldata_df[column_BEFL_maincrop].isin(early_maincrops),
+                conf.columns["class_declared"],
+            ] = early_maincrop_classname
+
     # For rows with no class, set to UNKNOWN
     parceldata_df.fillna(value={column_output_class: "UNKNOWN"}, inplace=True)
 
@@ -1219,7 +1303,7 @@ def prepare_input_latecrop(
     parceldata_df.loc[
         (parceldata_df[column_output_class] == "UNKNOWN")
         & (parceldata_df[column_BEFL_status_perm_grass] is not None)
-        & (parceldata_df["MON_CROPGROUP_MAINCROP"] == "MON_CG_GRASSEN"),
+        & (parceldata_df["MON_CG_MAINCROP"] == "MON_CG_GRASSEN"),
         column_output_class,
     ] = "MON_CG_GRASSEN"
     """
@@ -1229,7 +1313,7 @@ def prepare_input_latecrop(
     parceldata_df.loc[
         (parceldata_df[column_output_class] == "UNKNOWN")
         # & (parceldata_df[column_BEFL_status_perm_grass] is None)
-        & (parceldata_df["MON_CROPGROUP_MAINCROP"] != "MON_CG_GRASSEN")
+        & (parceldata_df["MON_CG_MAINCROP"] != "MON_CG_GRASSEN")
         & (parceldata_df["IS_PERM_BEDEKKING_MAINCROP"] == 1),
         column_output_class,
     ] = "IGNORE:PERMANENT_BEDEKKING"
@@ -1316,7 +1400,7 @@ def prepare_input_latecrop(
         parceldata_df.loc[
             (parceldata_df[ndvi_latecrop_median] <= 0.3)
             # & (parceldata_df[column_output_class] == "UNKNOWN")
-            & (parceldata_df["MON_CROPGROUP_MAINCROP"] != "MON_CG_GRASSEN"),
+            & (parceldata_df["MON_CG_MAINCROP"] != "MON_CG_GRASSEN"),
             column_output_class,
         ] = "MON_CG_BRAAK"
 
@@ -1354,57 +1438,6 @@ def prepare_input_latecrop(
             parceldata_df[column_BEFL_gesp_pm] = parceldata_df[
                 column_BEFL_gesp_pm
             ].str.replace(",", ";")
-
-    return parceldata_df
-
-
-def prepare_input_latecrop_early(
-    parceldata_df,
-    column_BEFL_latecrop: str,
-    column_BEFL_latecrop2: Optional[str],
-    column_BEFL_maincrop: str,
-    column_output_class: str,
-    classes_refe_path: Path,
-    min_parcels_in_class: int,
-    is_groundtruth: bool,
-):
-    """Prepare input file for use in the latecrop_early marker.
-
-    Prepare a classname column to classify the crop groups for early detection of
-    late crops. During this early detection, some main crops will still be on the field
-    during the entire perios. For these cases, replace the class with a class based on
-    the main crop.
-    """
-    # First run the standard latecrop prepare
-    parceldata_df = prepare_input_latecrop(
-        parceldata_df=parceldata_df,
-        column_BEFL_latecrop=column_BEFL_latecrop,
-        column_BEFL_latecrop2=column_BEFL_latecrop2,
-        column_BEFL_maincrop=column_BEFL_maincrop,
-        column_output_class=column_output_class,
-        classes_refe_path=classes_refe_path,
-        min_parcels_in_class=min_parcels_in_class,
-        is_groundtruth=is_groundtruth,
-    )
-
-    # Set parcels having a main crop that stays late on the field to another class, as
-    # the main crop will still be on the field:
-    # TODO: should be in REFE instead of hardcoded!!!
-    late_main_crops = {
-        "MON_CG_BIETEN": ["71", "91", "8532", "9532"],
-        "MON_CG_MAIS": ["201", "202"],
-        "IGNORE:LATE_MAINCROP": ["396"],
-    }
-
-    for class_name, cropcodes in late_main_crops.items():
-        parceldata_df.loc[
-            parceldata_df[column_BEFL_maincrop].isin(cropcodes), column_output_class
-        ] = class_name
-        if not is_groundtruth:
-            parceldata_df.loc[
-                parceldata_df[column_BEFL_maincrop].isin(cropcodes),
-                conf.columns["class_declared"],
-            ] = class_name
 
     return parceldata_df
 
@@ -1686,6 +1719,201 @@ def prepare_input_landcover_early(
     """Prepare input file for use in the landcover_early marker."""
     # First run the standard landcover prepare
     parceldata_df = prepare_input_landcover(
+        parceldata_df,
+        column_BEFL_cropcode,
+        column_output_class,
+        classes_refe_path,
+        min_parcels_in_class=min_parcels_in_class,
+        is_groundtruth=is_groundtruth,
+    )
+
+    # Set crops not in early crops to ignore
+    parceldata_df.loc[
+        parceldata_df[column_BEFL_earlylate] != "MON_TEELTEN_VROEGE",
+        column_output_class,
+    ] = "IGNORE:LATE_CROP"
+
+    # Set new grass to ignore
+    if column_BEFL_status_perm_grass in parceldata_df.columns:
+        parceldata_df.loc[
+            (parceldata_df[column_BEFL_cropcode] == "60")
+            & (
+                (parceldata_df[column_BEFL_status_perm_grass] == "BG1")
+                | (parceldata_df[column_BEFL_status_perm_grass].isnull())
+            ),
+            column_output_class,
+        ] = "IGNORE:NEW_GRASSLAND"
+    else:
+        logger.warning(
+            f"Source file doesn't contain column {column_BEFL_status_perm_grass}, so "
+            "new grassland cannot be ignored!"
+        )
+
+    return parceldata_df
+
+
+def prepare_input_ruggenteelt(
+    parceldata_df,
+    column_BEFL_cropcode: str,
+    column_output_class: str,
+    classes_refe_path: Path,
+    min_parcels_in_class: int,
+    is_groundtruth: bool,
+):
+    """Prepare input file for use in the ruggenteelt marker.
+
+    It should be a csv file with the following columns:
+        - object_id: column with a unique identifier
+        - classname: a string column with a readable name of the classes that will be
+          classified to
+
+    This specific implementation converts the typical export format used in BE-Flanders
+    to this format.
+    """
+    # Check if parameters are OK and init some extra params
+    # -----------------------------------------------------
+    if not classes_refe_path.exists():
+        raise Exception(f"Input classes file doesn't exist: {classes_refe_path}")
+
+    # Convert the crop to unicode, in case the input is int...
+    if column_BEFL_cropcode in parceldata_df.columns:
+        parceldata_df[column_BEFL_cropcode] = parceldata_df[
+            column_BEFL_cropcode
+        ].astype("string")
+
+    # Read and cleanup the mapping table from crop codes to classes
+    # -------------------------------------------------------------
+    logger.info(f"Read classes conversion table from {classes_refe_path}")
+    classes_df = pdh.read_file(classes_refe_path)
+
+    # Because the file was read as ansi, and gewas is int, so the data needs to be
+    # converted to unicode to be able to do comparisons with the other data
+    classes_df[column_BEFL_cropcode] = classes_df["CROPCODE"].astype("string")
+
+    # Map column MON_group to orig classname
+    column_output_class_orig = column_output_class + "_orig"
+    classes_df[column_output_class_orig] = classes_df[conf.columns["class_refe"]]
+
+    # Remove unneeded columns
+    for column in classes_df.columns:
+        if (
+            column not in [column_output_class_orig, column_BEFL_cropcode]
+            and column not in columns_BEFL_to_keep
+        ):
+            classes_df.drop(column, axis=1, inplace=True)
+
+    # Set the index
+    classes_df.set_index(column_BEFL_cropcode, inplace=True, verify_integrity=True)
+
+    # Get only the columns in the classes_df that don't exist yet in parceldata_df
+    cols_to_join = classes_df.columns.difference(parceldata_df.columns)
+
+    # Join/merge the classname
+    logger.info("Add the classes to the parceldata")
+    parceldata_df = parceldata_df.merge(
+        classes_df[cols_to_join],
+        how="left",
+        left_on=column_BEFL_cropcode,
+        right_index=True,
+        validate="many_to_one",
+    )
+
+    # Copy orig classname to classification classname
+    parceldata_df.insert(
+        loc=0, column=column_output_class, value=parceldata_df[column_output_class_orig]
+    )
+
+    # If a column with extra info exists, use it as well to fine-tune the classification
+    # classes.
+    if column_BEFL_gesp_pm in parceldata_df.columns:
+        # Serres, tijdelijke overkappingen en loodsen
+        parceldata_df.loc[
+            parceldata_df[column_BEFL_gesp_pm].isin(
+                ["SER", "PLA", "PLO", "SGM", "NPO", "LOO"]
+            ),
+            column_output_class,
+        ] = "MON_RUG_OVERK_LOO"
+        # Containers
+        parceldata_df.loc[
+            parceldata_df[column_BEFL_gesp_pm].isin(["CON"]), column_output_class
+        ] = "MON_RUG_CONTAINERS"
+        # TODO: CIV, containers in volle grond, lijkt niet zo specifiek te zijn...
+        # parceldata_df.loc[
+        #     parceldata_df[column_BEFL_gesp_pm] == "CIV", class_columnname
+        # ] = "MON_CONTAINERS"  # Containers, niet op volle grond...
+    else:
+        logger.warning(
+            f"The column {column_BEFL_gesp_pm} doesn't exist, so this part of the code "
+            "was skipped!"
+        )
+
+    # Some extra cleanup: classes starting with 'nvt' or empty ones
+    logger.info(
+        "Set classes that are still empty, not specific enough or that contain to "
+        "little values to 'UNKNOWN'"
+    )
+    parceldata_df.loc[
+        parceldata_df[column_output_class].str.startswith("nvt", na=True),
+        column_output_class,
+    ] = "MON_RUG_ONBEKEND_MET_KLASSIFICATIE"
+
+    # Set classes with very few elements to IGNORE:NOT_ENOUGH_SAMPLES!
+    if not is_groundtruth and min_parcels_in_class > 1:
+        for _, row in (
+            parceldata_df.groupby(column_output_class)
+            .size()
+            .reset_index(name="count")
+            .iterrows()
+        ):
+            if row["count"] < min_parcels_in_class:
+                logger.info(
+                    f"Class <{row[column_output_class]}> only contains {row['count']} "
+                    "elements, so put them to IGNORE:NOT_ENOUGH_SAMPLES"
+                )
+                parceldata_df.loc[
+                    parceldata_df[column_output_class] == row[column_output_class],
+                    column_output_class,
+                ] = "IGNORE:NOT_ENOUGH_SAMPLES"
+
+        # Add copy of class as class_declared
+        if conf.columns["class_declared"] not in parceldata_df.columns:
+            parceldata_df[conf.columns["class_declared"]] = parceldata_df[
+                column_output_class
+            ]
+
+    # Drop the columns that aren't useful at all
+    for column in parceldata_df.columns:
+        if (
+            column
+            not in [
+                column_output_class,
+                conf.columns["id"],
+                conf.columns["class_groundtruth"],
+                conf.columns["class_declared"],
+            ]
+            and column not in conf.preprocess.getlist("extra_export_columns")
+            and column not in columns_BEFL_to_keep
+        ):
+            parceldata_df.drop(column, axis=1, inplace=True)
+        elif column == column_BEFL_gesp_pm:
+            parceldata_df[column_BEFL_gesp_pm] = parceldata_df[
+                column_BEFL_gesp_pm
+            ].str.replace(",", ";")
+
+    return parceldata_df
+
+
+def prepare_input_ruggenteelt_early(
+    parceldata_df,
+    column_BEFL_cropcode: str,
+    column_output_class: str,
+    classes_refe_path: Path,
+    min_parcels_in_class: int,
+    is_groundtruth: bool,
+):
+    """Prepare input file for use in the ruggenteelt_early marker."""
+    # First run the standard ruggenteelt prepare
+    parceldata_df = prepare_input_ruggenteelt(
         parceldata_df,
         column_BEFL_cropcode,
         column_output_class,