+++ CHANGELOG.md
@@ -24,6 +24,7 @@
     - Remove `general.run_id`.
 - Rename some main modules to make naming more logical with non-cropclassification
   markers being added (#162)
+- Add RUGGENTEELT-EARLY and RUGGENTEELT detection (#187)
 
 ### Improvements
 
@@ -43,6 +44,8 @@
 - Add support to generate rvi index (#146)
 - Add support to generate vvdvh index (#151)
 - Add support to generate sarrgb images and apply enhanced lee despeckling (#157)
+- Add LATECROP-LATE detection (#177)
+- Add support for groundtruth data processing for latecrop marker (#193)
 - Add POC (not for operational use) of a cover/bare soil marker (#168)
 - General small improvements, e.g. save randomforest models compressed,.. (#144)
 - Use Exactextract as default engine for zonalstats calculation (#181)